% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.r
\name{checkSortingErrors}
\alias{checkSortingErrors}
\title{Sorting Error.}
\usage{
checkSortingErrors(estOrder, A)
}
\arguments{
\item{A:}{The true weighted adjacency matrix.}
}
\value{
Returns proportion of ancestors sorted after descendants if argument M is a DAG adjancency matrix, then returns proportion of parents sorted after children.
}
\description{
Sorting Error.
}
