% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.r
\name{getWeights}
\alias{getWeights}
\title{Get weighted adjancency matrix for linear SEM.
Assumes that X is zero-centered, which can be done via X=scale(Xoriginal,center=T,scale=F).
Estimation is done via Ordinary Least Squares (OLS) regression.}
\usage{
getWeights(X, pa)
}
\arguments{
\item{X}{An n by p data matrix which is used to estimate the weighted adjacency matrix.}

\item{pa}{A length p list object whose j-th entry gives the support for node j's parent set.}
}
\value{
A p by p weighted adjacency matrix for the linear SEM.
}
\description{
Get weighted adjancency matrix for linear SEM.
Assumes that X is zero-centered, which can be done via X=scale(Xoriginal,center=T,scale=F).
Estimation is done via Ordinary Least Squares (OLS) regression.
}
