% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.r
\name{moralize}
\alias{moralize}
\title{Moralize a DAG. 
Obtain Markov Blanket, the union of all children, parents, co-parents.}
\usage{
moralize(B)
}
\arguments{
\item{B}{The p by p weighted adjacency matrix. Assumes B[j,k]!=0 if \eqn{j\in PA_k}.}
}
\value{
A list of length p. Each entry j corresponds to the Markov Blanket of node j according to adjacency matrix B.
}
\description{
Moralize a DAG. 
Obtain Markov Blanket, the union of all children, parents, co-parents.
}
