% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.r
\name{randomWeightedAdjacencyMatrix}
\alias{randomWeightedAdjacencyMatrix}
\title{Generate a random weighted adjacency matrix with num.roots number of root nodes.
Each non-root node will have between pa.min and pa.max number of parents.
The parent weight (in absolute value) is between pa.wt.min and pa.wt.max.
The parent weight is positive with probability prob.pos (else, negative).}
\usage{
randomWeightedAdjacencyMatrix(
  p,
  num.roots = p - 1,
  pa.min = 1,
  pa.max = 1,
  pa.wt.min = 1,
  pa.wt.max = 1,
  prob.pos = 0.5,
  perm = sample(x = p, size = p, replace = F)
)
}
\arguments{
\item{p}{Number of nodes in underlying DAG.}

\item{num.roots}{Number of root nodes in underlying DAG. Default: num.roots=p-1.}

\item{pa.min}{Minimum number of parents a node can have. Default: pa.min=1.}

\item{pa.max}{Maximum number of parents a node can have. Default: pa.max=1.}

\item{pa.wt.min}{The minimum (in absolute value) coefficient value for the parent to a child in the linear SEM. Default: pa.wt.min=1.}

\item{pa.wt.max}{The maximum (in absolute value) coefficient value for the parent to a child in the linear SEM. Default: pa.wt.max=1.}

\item{prob.pos}{The probability of a positive coefficient value for the parent to a child in the the linear SEM. Default: prob.pos=0.5.}

\item{perm}{A topological ordering for underlying DAG. Default is random.}
}
\value{
A list with two elements: perm (the topological ordering used) and B (the weighted adjacency matrix).
}
\description{
Generate a random weighted adjacency matrix with num.roots number of root nodes.
Each non-root node will have between pa.min and pa.max number of parents.
The parent weight (in absolute value) is between pa.wt.min and pa.wt.max.
The parent weight is positive with probability prob.pos (else, negative).
}
