% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.r
\name{skeleton}
\alias{skeleton}
\title{Adjacency matrix for an undirected (moral) graph based on Markov Blanket (or neighborhood estimates) of each node.}
\usage{
skeleton(mb)
}
\arguments{
\item{mb}{The Markov blanket or neighborhood estiamtes. A list of length p.}
}
\value{
A p by p adjacency matrix for an undirected graph. 0 indicates no undirected edge in the graph, while 1 indicates an directed edge in the graph.
}
\description{
Adjacency matrix for an undirected (moral) graph based on Markov Blanket (or neighborhood estimates) of each node.
}
