% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.r
\name{checkSortingErrors}
\alias{checkSortingErrors}
\title{Sorting Error.}
\usage{
checkSortingErrors(estOrder, A)
}
\arguments{
\item{estOrder:}{A length p vector with the estimated topological ordering. Its unique elements are a re-ordering of the indices 1,2,...,p.}

\item{A:}{The true DAG adjacency matrix.}
}
\value{
Returns proportion of parents sorted after children. See Equation (11) in the paper, which defines the output of this function: \eqn{{\rm ERR}(\hat{\pi})}.
}
\description{
Sorting Error.
}
