% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.r
\name{generateData}
\alias{generateData}
\title{Generate Linear (laplace, logistic, or scaled-t error family) Structural Equation Model Data}
\usage{
generateData(
  B,
  n = 1000,
  family = "laplace",
  scale = rep(1, ncol(B)),
  perm = topoSort(as.matrix(B)),
  df = 5
)
}
\arguments{
\item{B}{A p by p weighted adjacency matrix for an acyclic linear SEM.}

\item{n}{The desired sample size. Default: n=1000.}

\item{family}{Currently allow 't', 'laplace' (default), or 'logistic'.}

\item{df}{Degrees of freedom for scaled-t-distributed noise. Default: df=5.}
}
\value{
An n by p matrix object.
}
\description{
Generate Linear (laplace, logistic, or scaled-t error family) Structural Equation Model Data
}
