\name{scorelingam-package}
\alias{scorelingam-package}
\alias{scorelingam}
\docType{package}
\title{
\packageTitle{scorelingam}
}
\description{
\packageDescription{scorelingam}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scorelingam}
\packageIndices{scorelingam}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{scorelingam}

Maintainer: \packageMaintainer{scorelingam}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
