\name{scorelingam-package}
\alias{scorelingam-package}
\alias{scorelingam}
\docType{package}
\title{
  ScoreLiNGAM
}
\description{
  Implements the sorting procedure outlined in \href{https://openreview.net/forum?id=4pCjIGIjrt}{Ruiz et. al (2022+)}. 
}
\details{
  The main procedure is `scorelingam`. See its documentation below for more info. See the examples at https://github.com/gabriel-ruiz/scorelingam.
}
\author{
Gabriel Ruiz
}



\references{
G Ruiz, OH Madrid-Padilla, Q Zhou. "Sequentially learning the topological ordering of directed acyclic graphs with likelihood ratio scores." Transactions of Machine Learning Research, 2022+. [\href{https://openreview.net/forum?id=4pCjIGIjrt}{Link}]  
}
\keyword{ package }

\examples{
 See the examples at https://github.com/gabriel-ruiz/scorelingam.
}
